+++ src/options.h
@@ -14,7 +14,7 @@ enum class Action : int {
 // Use many magic numbers here. Just for convinience.
 struct Options {
     int poll_event_timeout_ms = -1;
-    int scroll_rows_per_mouse_wheel_scroll = 3;
+    int64_t scroll_rows_per_mouse_wheel_scroll = 3;
     std::vector<const char*> begin_files;
     // // Now only Support single Keys
     // std::unordered_map<Terminal::KeyInfo, Action> keymaps;
